chore: correct backend structure, remove unused frontend code and add font
