--- Models/UserModel.cs
@@ -1,9 +1,0 @@
-﻿namespace CapstoneProject_RiskExposureTrackingAndReportingSystem.Models
-{
-    public class UserModel
-    {
-        public string UserName { get; set; }
-        public string Password { get; set; }
-        public string? Role { get; set; }
-    }
-}
