+++ package.json
@@ -3,6 +3,8 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
+    "start": "concurrently \"npm run start --workspace=api\" \"npm run start --workspace=client\"",
+    "build": "concurrently \"npm run build --workspace=api\" \"npm run build --workspace=client\"",
     "dev": "concurrently \"npm run dev --workspace=api\" \"npm run dev --workspace=client\""
   },
   "workspaces": [
